backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
Pillow==10.4.0
django-countries==7.6.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
